Remove unnecessary fragment
