+++ playground/src/main/java/mx/yellowme/youst/playground/ui/chart/ChartActivity.kt
@@ -1,11 +1,17 @@
 package mx.yellowme.youst.playground.ui.chart
 
 import android.os.Bundle
+import android.util.Log
 import kotlinx.android.synthetic.main.screen_chart.*
 import mx.yellowme.youst.core.hooks.BaseActivity
+import mx.yellowme.youst.core.utils.loadJsonObject
 import mx.yellowme.youst.playground.R
-import mx.yellowme.youst.playground.components.AppHeroActionListener
-import mx.yellowme.youst.playground.ui.chart.screens.ChartViewFragment
+import mx.yellowme.youst.playground.components.OnChangeListener
+import mx.yellowme.youst.playground.data.ChartFakeRepository
+import mx.yellowme.youst.playground.data.SingleItemCallback
+import mx.yellowme.youst.playground.domain.ChartEntry
+import mx.yellowme.youst.playground.domain.ChartSetting
+import mx.yellowme.youst.playground.domain.ChartType
 
 /**
  * @author adrianleyvasanchez
@@ -13,9 +19,11 @@ import mx.yellowme.youst.playground.ui.chart.screens.ChartViewFragment
  */
 class ChartActivity : BaseActivity() {
 
-    //Region attributes
+    //Region Attributes
 
     override val layoutId = R.layout.screen_chart
+
+    private val settingsJsonPath = "chart_settings.json"
 
     //endregion
 
@@ -23,17 +31,45 @@ class ChartActivity : BaseActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        appHero.run {
-            message = getString(R.string.screen_chart)
-            mainActionListener = object : AppHeroActionListener {
-                override fun onClickAction() {
-                    supportFragmentManager
-                        .beginTransaction()
-                        .replace(android.R.id.content, ChartViewFragment())
-                        .commit()
+        loadJsonObject<ChartSetting>(settingsJsonPath)?.let {
+            yellowChart?.run {
+                label = getString(R.string.showcases_label)
+                chartSettings = it
+            }
+
+            chartSelector?.run {
+                currentTypeSelected = ChartType.valueOf(it.type)
+                delegate = object : OnChangeListener {
+                    override fun didChangeChartType(type: ChartType) {
+                        yellowChart?.updateType(type.toString())
+                    }
                 }
             }
         }
+
+        repeat(20) {
+            getData()
+        }
+    }
+
+    //endregion
+
+    //region Helpers
+
+    //TODO: Must delegate action to another layer component (ViewModel or Presenter)
+    private fun getData() {
+        ChartFakeRepository().getData(false,
+            object : SingleItemCallback<ChartEntry> {
+                override fun onServerError(message: String) {
+                    Log.d("tag", message)
+                }
+
+                override fun onLoad(item: ChartEntry?) {
+                    item?.let {
+                        yellowChart?.addData(item)
+                    }
+                }
+            })
     }
 
     //endregion
