--- playground/src/main/res/layout/screen_chart_view.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/white"
-    android:orientation="vertical">
-
-    <mx.yellowme.youst.playground.components.YellowChart
-        android:id="@+id/yellowChart"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight=".85"
-        android:background="@color/white"
-        android:gravity="center"
-        android:orientation="horizontal"
-        android:paddingStart="@dimen/dimen_5"
-        android:paddingTop="@dimen/dimen_40"
-        android:paddingEnd="@dimen/dimen_5"
-        android:paddingBottom="@dimen/dimen_20" />
-
-    <mx.yellowme.youst.playground.components.ChartSelector
-        android:id="@+id/chartSelector"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight=".15"
-        app:selectorBackgroundColor="@color/design_default_color_primary"
-        app:selectorOptionsTextColor="@color/white"
-        app:selectorOptionsHighlightColor="@color/white" />
-
-</LinearLayout>
